Added typescript and refactored code
