+++ 01-Github-Explorer/babel.config.js
@@ -1,6 +1,7 @@
 module.exports = {
     presets: [
         '@babel/preset-env',
+        '@babel/preset-typescript',
         ['@babel/preset-react', {
             runtime: 'automatic'
         }]
