+++ 01-Github-Explorer/package.json
@@ -16,7 +16,11 @@
     "@babel/core": "^7.17.9",
     "@babel/preset-env": "^7.16.11",
     "@babel/preset-react": "^7.16.7",
+    "@babel/preset-typescript": "^7.16.7",
     "@pmmmwh/react-refresh-webpack-plugin": "^0.5.5",
+    "@types/react": "^18.0.5",
+    "@types/react-dom": "^18.0.1",
+    "@types/typescript": "^2.0.0",
     "babel-loader": "^8.2.4",
     "cross-env": "^7.0.3",
     "css-loader": "^6.7.1",
@@ -25,6 +29,7 @@
     "react-refresh": "^0.12.0",
     "sass-loader": "^12.6.0",
     "style-loader": "^3.3.1",
+    "typescript": "^4.6.3",
     "webpack": "^5.72.0",
     "webpack-cli": "^4.9.2",
     "webpack-dev-server": "^4.8.1"
