+++ 01-Github-Explorer/src/components/RepositoryItem.tsx
@@ -1,4 +1,12 @@
-export function RepositoryItem(props) {
+interface RepositoryItemprops {
+    repository: {
+        full_name: string,
+        description: string,
+        html_url: string
+    }
+}
+
+export function RepositoryItem(props: RepositoryItemprops) {
     return (
         <li>
             <strong>{props.repository.full_name ?? 'Default'}</strong>
