+++ 01-Github-Explorer/src/components/RepositoryList.tsx
@@ -2,8 +2,15 @@ import { RepositoryItem } from "./RepositoryItem";
 import '../styles/repositories.scss';
 import { useState, useEffect } from "react";
 
+interface Repository {
+    id: string,
+    full_name: string,
+    description: string,
+    html_url: string
+}
+
 export function RepositoryList() {
-    const [repositories, setRepositories] = useState([]);
+    const [repositories, setRepositories] = useState<Repository[]>([]);
 
     useEffect(() => {
         fetch('https://api.github.com/orgs/rocketseat/repos')
