+++ 01-Github-Explorer/src/index.tsx
@@ -1,4 +1,4 @@
 import { render } from 'react-dom';
-import { App } from './app';
+import { App } from './App';
 
 render(<App/>, document.getElementById('root'));
