+++ 01-Github-Explorer/tsconfig.json
@@ -1,0 +1,19 @@
+{
+  "compilerOptions": {
+    "target": "es2016",                                  
+    "lib": ["dom", "dom.iterable", "esnext"],                                        
+    "jsx": "react-jsx",                                
+    "module": "commonjs",                                
+    "moduleResolution": "node",    
+    "resolveJsonModule": true,
+    "isolatedModules": true,                   
+    "allowJs": true,                                  
+    "noEmit": true,                                   
+    "allowSyntheticDefaultImports": true,             
+    "esModuleInterop": true,                             
+    "forceConsistentCasingInFileNames": true,
+    "strict": true,                                      
+    "skipLibCheck": true                                 
+  },
+  "include": ["src"]
+}
